refactor: use /src structure to clarify what the service exports
Related to #313
