+++ source-code/website/src/services/markdown/index.ts
@@ -5,8 +5,8 @@
  * --------------------------------
  */
 
-export { parseMarkdown, RequiredFrontmatter } from "./parseMarkdown.js";
-export { Markdown } from "./Markdown.jsx";
+export { parseMarkdown, RequiredFrontmatter } from "./src/parseMarkdown.js";
+export { Markdown } from "./src/Markdown.jsx";
 export {
   /**
    * Use z(od) to extend the required frontmatter schema.
